Now it works.

+++ Server/packet.cpp
@@ -8,8 +8,8 @@ void sendPacket(char type, char data1, char data2, int server) {
     char buffer[3];
     
     buffer[0] = type;
-    buffer[1] = data1;
-    buffer[2] = data2;
+    buffer[1] = (int) data1;
+    buffer[2] = (int) data2;
 
     send(server, buffer, sizeof(char) * 3, 0);
 }
@@ -17,7 +17,7 @@ void sendPacket(char type, char data1, char data2, int server) {
 Packet receivePacket(char buffer[3]) {
     Packet packet;
 
-    packet.type = buffer[0];
+    packet.type  = buffer[0];
     packet.data1 = buffer[1];
     packet.data2 = buffer[2];
